+++ src/ShaderPlayground.Core/LanguageNames.cs
@@ -12,6 +12,7 @@
         public const string Ispc = "ISPC";
         public const string Slang = "Slang";
         public const string Smolv = "SMOL-V";
+        public const string Zlib = "zlib";
         public const string Zstd = "zst";
     }
 }